backend/requirements.txt
# === CORE FRAMEWORK & UTILITIES ===
fastapi==0.124.4
uvicorn[standard]==0.38.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.21
httpx==0.28.1
python-dotenv==1.2.1
email-validator==2.3.0
requests==2.32.5
nltk==3.9.2

# === AUTHENTICATION & GOOGLE INTEGRATION ===
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.3.0
wikipedia==1.4.0

# === DATABASE / SEARCH / NLP ===
faiss-cpu==1.13.1   # or faiss-gpu if you have CUDA
datasets==4.4.1
sentence-transformers==5.2.0

# === CORE AI MODELS (No External APIs!) ===
torch==2.9.1
transformers==4.57.3
accelerate==1.12.0
safetensors==0.7.0
bitsandbytes==0.49.0   # For model quantization (GPU)

# === VISION MODELS ===
timm==1.0.22            # Image models backbone
pillow==12.0.0
easyocr==1.7.2          # Better OCR than Tesseract
pytesseract==0.3.13     # Backup OCR

# === DOCUMENT HANDLING ===
pypdf==6.4.2
python-docx==1.2.0

# === VOICE (Optional) ===
SpeechRecognition==3.14.4
pyttsx3==2.99

# === ADDITIONAL ===
numpy==2.2.6
scikit-learn==1.8.0